Update optimizer, and rename to InferRules
